requirements.txt
Django==4.0.4
djangorestframework==3.13.1
Pillow==9.1.1
psycopg2-binary==2.9.3
django-environ
django-filter
django-imagekit